refactor(#step2): 메뉴 상품 수량 관련 VO 제거
